Refactored token namespace to SkoleProtokolLibrary.Authentication
